+++ src/MotionDriver/c_interpollation_hardware.h
@@ -28,7 +28,7 @@ namespace Motion_Core
 			static uint8_t Step_Active;
 			static uint8_t Interpolation_Active;
 			static uint32_t Current_Line;
-			static uint8_t Step_Pulse_Length;  // Step pulse reset time after step rise
+			static uint32_t Step_Pulse_Length;  // Step pulse reset time after step rise
 			static void step_tick();
 			static uint8_t is_active();
 			static void Initialize();
